# [Xamarin.Android.Build.Tasks] implement a MemoryStreamPool