--- a/src/Xamarin.Android.Build.Tasks/Utilities/XDocumentExtensions.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/XDocumentExtensions.cs
@@ -58,12 +58,11 @@ namespace Xamarin.Android.Tasks
 
 		public static bool SaveIfChanged (this XDocument document, string fileName)
 		{
-			using (var stream = new MemoryStream ())
-			using (var sw = new StreamWriter (stream))
+			using (var sw = MemoryStreamPool.Shared.CreateStreamWriter ())
 			using (var xw = new Monodroid.LinePreservedXmlWriter (sw)) {
 				xw.WriteNode (document.CreateNavigator (), false);
 				xw.Flush ();
-				return MonoAndroidHelper.CopyIfStreamChanged (stream, fileName);
+				return MonoAndroidHelper.CopyIfStreamChanged (sw.BaseStream, fileName);
 			}
 		}
 	}